+++ include/lipm_walking/ModelPredictiveControl.h
@@ -35,7 +35,7 @@
 #include <lipm_walking/Contact.h>
 #include <lipm_walking/Pendulum.h>
 #include <lipm_walking/Preview.h>
-#include <lipm_walking/defs.h>
+#include <lipm_walking/utils/world.h>
 
 namespace lipm_walking
 {