+++ include/lipm_walking/Preview.h
@@ -28,7 +28,6 @@
 #pragma once
 
 #include <lipm_walking/Pendulum.h>
-#include <lipm_walking/defs.h>
 
 namespace lipm_walking
 {