+++ include/lipm_walking/utils/world.h
@@ -29,15 +29,10 @@
 
 #include <Eigen/Dense>
 
-namespace lipm_walking
+namespace world
 {
-  namespace world
-  {
-    constexpr double GRAVITY = 9.80665; // ISO 80000-3
+  constexpr double GRAVITY = 9.80665; // ISO 80000-3
 
-    const Eigen::Vector3d e_z = Eigen::Vector3d{0., 0., 1.};
-    const Eigen::Vector3d gravity = Eigen::Vector3d{0., 0., -GRAVITY};
-  }
-
-  const Eigen::Vector3d e_z = {0., 0., 1.};
+  const Eigen::Vector3d gravity = Eigen::Vector3d{0., 0., -GRAVITY};
+  const Eigen::Vector3d vertical = Eigen::Vector3d{0., 0., 1.};
 }